api refactor

+++ server/src/services/branch.service.ts
@@ -1,31 +1,15 @@
 import { prisma } from '../config/prisma';
 import { HttpException } from '../exceptions/httpException';
 
-const getBranches = async () => {
-  const branches = await prisma.branch.findMany({
-    select: {
-      id: true,
-      title: true,
-      address: true,
-    },
-  });
+const getCorporationBranches = async (corporationId: string) => {
+  const corporation = await prisma.corporation.findUnique({ where: { id: corporationId } });
 
-  return branches;
-};
-
-const getBranch = async (id: string) => {
-  const branch = await prisma.branch.findUnique({ where: { id } });
-
-  if (!branch) {
-    throw new HttpException(404, 'Branch not found!');
+  if (!corporation) {
+    throw new HttpException(404, 'Not found!');
   }
 
-  return branch;
-};
-
-const getCorporationBranches = async (id: string) => {
   const corporationBranches = await prisma.branch.findMany({
-    where: { corporationId: id },
+    where: { corporationId: corporationId },
     select: {
       id: true,
       title: true,
@@ -36,15 +20,35 @@ const getCorporationBranches = async (id: string) => {
   return corporationBranches;
 };
 
-const createBranch = async (data: { title: string; address: string; corporationId: string }) => {
+const getBranch = async (corporationId: string, id: string) => {
+  const branch = await prisma.branch.findFirst({ where: { id, corporationId } });
+
+  if (!branch) {
+    throw new HttpException(404, 'Not found!');
+  }
+
+  return branch;
+};
+
+const createBranch = async (data: { title: string; address: string }, corporationId: string) => {
   const branch = await prisma.branch.create({
-    data: { title: data.title, address: data.address, corporationId: data.corporationId },
+    data: { title: data.title, address: data.address, corporationId: corporationId },
   });
 
   return branch;
 };
 
-const updateBranch = async (data: { title: string; address: string }, id: string) => {
+const updateBranch = async (
+  data: { title: string; address: string },
+  corporationId: string,
+  id: string
+) => {
+  const tempBranch = await prisma.branch.findFirst({ where: { id, corporationId } });
+
+  if (!tempBranch) {
+    throw new HttpException(404, 'Not found!');
+  }
+
   const branch = await prisma.branch.update({
     where: { id: id },
     data: { title: data.title, address: data.address },
@@ -53,7 +57,13 @@ const updateBranch = async (data: { title: string; address: string }, id: string
   return branch;
 };
 
-const deleteBranch = async (id: string) => {
+const deleteBranch = async (corporationId: string, id: string) => {
+  const tempBranch = await prisma.branch.findFirst({ where: { id, corporationId } });
+
+  if (!tempBranch) {
+    throw new HttpException(404, 'Not found!');
+  }
+
   const branch = await prisma.branch.delete({
     where: { id: id },
   });
@@ -62,7 +72,6 @@ const deleteBranch = async (id: string) => {
 };
 
 export const branchService = {
-  getBranches,
   getBranch,
   getCorporationBranches,
   createBranch,
